+++ source/extensions/omni.isaac.isaac_sensor/python/scripts/samples/imu_sensor.py
@@ -20,7 +20,7 @@ from pxr import Gf, UsdGeom
 
 from omni.isaac.ui.ui_utils import setup_ui_headers, get_style, LABEL_WIDTH
 from omni.kit.menu.utils import add_menu_items, remove_menu_items, MenuItemDescription
-
+from omni.isaac.core.utils.nucleus import get_assets_root_path
 
 EXTENSION_NAME = "IMU Sensor Example"
 
@@ -178,16 +178,20 @@ class Imu_sensor_demo(omni.ext.IExt):
                 self.sliders[9].model.set_value(1)
 
     async def create_scenario(self):
+        self._assets_root_path = get_assets_root_path()
+        if self._assets_root_path is None:
+            carb.log_error("Could not find Isaac Sim assets folder")
+            return
 
         # Add IMU Sensor
-        await omni.usd.get_context().open_stage_async(self._extension_path + "/data/ant.usd")
+        await omni.usd.get_context().open_stage_async(self._assets_root_path + "/Isaac/Robots/Simple/ant.usd")
         await omni.kit.app.get_app().next_update_async()
 
         self.meters_per_unit = UsdGeom.GetStageMetersPerUnit(omni.usd.get_context().get_stage())
 
         result, sensor = omni.kit.commands.execute(
             "IsaacSensorCreateImuSensor",
-            path="/sensor",
+            path="/ant/sensor",
             parent=self.body_path,
             sensor_period=1 / 500.0,
             translation=Gf.Vec3d(0, 0, 0),